tarefa jdbc

+++ zenon-fraud-detector/src/main/java/br/com/zenon/Transaction.java
@@ -3,7 +3,7 @@ package br.com.zenon;
 import java.math.BigDecimal;
 import java.util.Objects;
 
-public record Transaction(int step, TransactionType type, BigDecimal amount, TransactionCustomer origin,
+public record Transaction(long step, TransactionType type, BigDecimal amount, TransactionCustomer origin,
                           TransactionCustomer recipient, boolean isFraud, boolean isFlaggedFraud) {
     public Transaction {
         Objects.requireNonNull(type);
